Move code from handlers.js into serviceFacades/instantlaunches
